vectors: AES-PMAC-SIV test vectors
As AES-PMAC-SIV is effectively a novel construction, it has no existing test
vectors to draw from.

This set of test vectors was created by using the same inputs as the AES-SIV
test vectors, and running them against the Rust AES-PMAC-SIV implementation.

This implementation should be correct because:

1) AES-PMAC has been tested against Rogaway's test vectors
2) AES-SIV has been tested against several vectors
3) AES-SIV is implemented as a generic construction, letting us swap within
   a known-good AES-SIV implementation a known-good PMAC implementation.

Based on this, we use the Rust implementation as a reference, and generate test
vectors from it in order to test correctness of the other language
implementations.