go: PMAC implementation
This adds an implementation of PMAC which provides the `hash.Hash` interface.

The implementation has been tested against the common set of AES-PMAC test
vectors and passes.

This implementation presently is not vectorized. It looks like Go does not make
it simple to use vectorized AES-NI, and vectorizing it would likely require
using an ASM implementation instead of `cipher.Block`.