[WIP] rust: KDF + STREAM support
Presently hidden under the "yolocrypto" flag.

Please note that "yolocrypto" should not be used for any purpose
whatsoever.

This branch contains research and experimentation towards what a
minimal, parsimonious STREAM implementation would look like.

Part of that is using a KDF to derive a unique key per STREAM.
This PR includes a "placeholder" KDF for prototyping purposes which is
intended to be replaced with a sounder, more standard cipher-based KDF.