README.md: Remove SparklingHeart!™ ratings
I tried adding a fourth tier in my heart powered gumball chart-style
competitive analysis consisting of red, yellow, and green hearts, adding
a SparklingHeart!™ for maximal awesomeness.

The SparklingHeart!™ unfortunately renders oddly on some platforms and
may be overselling the things labeled with it (or underselling the
things not labeled with it) and, in retrospect, just makes the chart
more complicated and confusing (also resorting in ♣ as a typographic
dagger).

The resulting chart should be simpler, easier to interpret, and
hopefully more fair to every algorithm listed.