Revert "rust: Add Buffer type"
This reverts commit 901690b203446d59c32b93aad9797de75e916ab1.

The Buffer type was a neat idea, but it has a few drawbacks I didn't
notice initially which I think present significant complications.

The main one is that it relies on its contents being

`T: AsRef<[u8]> + AsMut<[u8]>`

Unfortunately until const generics land, these traits are only defined
for up to 32-byte arrays! This makes the in-place API unusable for
larger fixed-size arrays.

We can revisit this when const generics land.

The other problem is it requires an owned type for any data passed to
it, which may not always be the case. This seems like a larger
dealbreaker.

The slice-based API is ugly but certainly the most flexible.